Made DatePicker
Created an HTML datepicker with custom calnder icon instead of arrow. Disabled keyboard editing of date, so user is forced to use the calender icon and can't enter a date below the min or above the max.

Created a JS function that sets the date range's max value to the current date.