Date Range Functionality
Created the datepickers in HTML5 which work in Chrome but not Firefox. Tweaked the design, made them only changeable through calendar icon, set start and end dates.

Added temp fix to display an alternative datepicker in a browser that doesn't support html5 date type using webshims.

Probably should just change the html5 datepicker to JQuery DatePicker so that the style is consistant.